Move to using PEG completely, added comments to parser.
Makefile was modified a bit to take out the rule for `final_branch`.
Didn't seem like that was going to work out, no need to have it dangling
there.

Also fixed a bug in the parser where mapping a single signal to a single
signal didn't cause an actual pin to appear. Also added support for
comments in the pinfiles.

Changed pinfile_builder.sh to use PEG based parser.

 Changes to be committed:
	modified:   Makefile
	modified:   peg_parsing.js
	modified:   pegjs_parser.pegjs
	renamed:    peg_pinfileBuilder.js -> pinfileBuilder.js
	modified:   pinfile_builder.sh